Add files via upload
Complex network topology is created to configure BGP, EIGRP and OSPF. The border routers are configured with BGP and two virtual sections are configured with OSPF and EIGRP. No route redistribution is performed, and hence the topology fails. The aim of this section was to show the importance of route redistribution.